+++ Lecon/requirements.txt
@@ -1,0 +1,3 @@
+SQLAlchemy==2.0.0
+psycopg2-binary==2.9.3
+numpy==1.23.4
